+++ Campaign/config/ziggy.php
@@ -2,5 +2,5 @@
 
 return [
     // Filter laravel paths that are passed to JS via Ziggy library
-    'whitelist' => ['comparison.*'],
+    'only' => ['comparison.*'],
 ];