+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - **BREAKING**: Upgraded to Laravel 7. remp/remp#491
   - All API endpoints now return dates in ISO-8601 compatible format. Make sure all code consuming the Campaign API is accustomed to this change. For example, previously, date would serialized like the following: `2019-12-02 20:01:00`. Now it is serialized like `2019-12-02T20:01:00.283041Z` (always in UTC).
   - Environment variable `APP_SESSION_EXPIRATION` was renamed to `SESSION_LIFETIME`.
+  - Fixed broken campaign comparison. remp/remp#926
+  - Fixed broken campaign copying. remp/remp#927
 
 ### [Sso]